Update ProxyManager to use UUID for instance identification.
When it starts a new instance, it needs to confirm that it's successfully
started, to do this, we need to check implicit relation with IPC connection
and the instance.

And also we can't sure instances are separated by process. We should
use instance manager for stopping instance when it sees timeout in
connection.