Handle execution environment failure of proxy
If system meets execution environment failure, it needs to try
it again. In other critical cases, it should return it to the system
for stopping the chain.

For this, it changes result delivery system. Now, it uses error object
for it, and error code areas are changed according to this.
And also lots of errors are adjusted according to the change.