Update SM for handling null value
Previously, we doesn't assume that the contract supply nil values in
events. But it's allowed in Core1, so we decide to follow the policy.
But in Core1, it does accept nil value for required arguments, but
we will not follow the policy. It accepts nil only if default value
is nil.

- Allow events with nil values
- Allow to supply required arguments only if others are optional
- Allow to give nil if and only if default value is nil
- Update test cases and SCOREs in testsuite
- Update structure for client to show null value