Performance tuning in world state
Let it stores key value of the account for reducing calculation
of sha3sum256. cacheID is redundant identifier. So, this patch
let cache use key instead of cacheID(same as address ID). And also
stores different field useCache as false instead of nil cacheID.

It reduces lots of calculation in making snapshot of world state.