Fix bug, invalid algorithm to check changes of storage
StorageChangeAfter() method is used to check whether it's changed
after the specified snapshot. But it returns whether it's same or
not (exactly opposite value).
It causes that the SCORE refers invalid system parameter while it
processes transactions in the same block.

In most cases in the enterprise environment, system parameter changes
are triggerd system management period. And in the period, normal
user transactions are not enqueued. Even if they did, it doesn't
do cause critical service failure.

But in public environment, it could cause lots of critical problems.