Optimize on allocating bytes in ompt
While it handles the request with the given key, it uses a slice
of bytes containing nibbles. Previous codes allocate it every time.
New code uses cached buffer. It allocates a new slice if it needs
to store it.

It reduces heap allocation and also enhances performance of the node.