Fee sharing v2.1
To enable fee sharing feature, "feeSharingEnabled" value of
the chain must be non-zero value.

For v2.0, "depositTerm" must be set as proper height of blocks.
Issue rate was fixed as 8%, but it can be configured by setting
"depositIssueRate" value. Minimum deposit amount is fixed as
5000e+18.

For v2.1, "depositTerm" must be set as zero. Total amount of the
deposit should exceeds fee limit ( price x step limit for invoke )
to pay the fee of users.