Improve transaction addition runs in more parallel way
Let it verify some information before adding the transaction to the
pool. Most handlers run in goroutine. Signature verification takes
more time than others. If it's handled in goroutine, we can accept
more transactions from the user.

It's not confirmed that whether it enhances actual performance or
not, but it's a harmless change system way.

To improve more, it needs to apply asynchronous logics in transaction
propagation handlers.