Send block precommit if we have a block and polka for it.
If a consensus engine have a complete block and polka for it but it had
not been imported, the engine sent nil precommit. Now, a consensus
engine sends block precommit in such case regardless if it had been
imported or not. It is OK to send block precommit in the case, even if
the import for the block would fail because the engine has polka which
means at least honest +1/3 nodes successfully imported the block before
they send prevote. Thus, the engine would be able to do forced sync in
commit step.