Update for vote verification failure issue.
Problem case: OpenNetwork -> CloseNetwork -> RevokeValidator ->
OpenNetwork

If all networks in a NT are closed, the NTS becomes inactive. Now,
Inactive NT's PC is not included in PCM. Later if a new network is open,
the first block is not voted since previous main block's PCM do not have
the PC for the NT. Then PCM is updated since the first block's
npcChanged flag is 1.