Fix problem in deposit management
It fixes incorrect calculation, but in mainnet, it doesn't happen.
In depositList implementation, it checks available deposit before
consuming deposits. So, There should be multiple deposits in the list
for the "amount" to exceeds "DepositRemain".
So, V1 and V2 deposits must coexist for the case.
But before executing the transaction it checks whether it's possible
to pay or not through "CanPay" method. And it checks for it for each
deposit. So, one of deposit should be possible to pay all.

If V1 deposit is possible to pay, then V2 will pay nothing.
If V2 deposit is possible to pay, then "DepositRemain" is greater than
equal to "amount". The only case to pass the changed statement is
that "DepositRemain" and "amount" have same values.
And in the case, result is same as previous codes.

Anyway, the logic is incorrect, then it should be fixed.