Improve test: do not remove TX when it is proposed in test SM
A TX in pool was removed if it was proposed in test SM. This made some
problem in some tests. Some tests send TX and wait for the TX to be
included in a block.  If a validator included the TX in a block proposal
and the proposal was not committed, the TX was never included in the
following proposal blocks of the validator. To avoid the case, remove TX
in pool only if it is committed in test SM.