Fix chain: verify votes for target block in reset task
After reset, the state will be valid if the trusted block hash is valid
as the operation resets database with trusted block hash and does not
use votes to verify block. However, theoretically it is possible for a
fooled node to propose a block with invalid previous votes. Fix the
behavior by validating vote during reset.