Fix javaee: crash due to wrong use of AdviceAdapter
The AdviceAdapter requires that all related instructions are propagated
to this adapter. Currently ArraysWithKnownTypesAdapter doesn't meet
this requirement. For example, it propagates visitVarInsn but it doesn't
propagate visitMethodInsn to AdviceAdapter. If this requirement is not
met, the frame stack inside AdviceAdapter becomes imbalanced and the
imbalance causes an exception. Since it does not use the function of
AdviceAdapter, fix the issue by just extending GeneratorAdaptor.